Add some config options
These are some of the different things we will need for querying
musicbrainz. The app_name and contact_info will be used for the
User-Agent header that is set on the requests, for if the user needs to
be contact by musicbrainz about usage.

The musicbrainz_url is just the base url for the search API.