Add functions for building Artist
The Artist module now has its own notion of how to build an Artist
struct from an input hash. For now it is almost a straight-across
mapping, but eventually there might be some type coercions as well, as
the idea is that the input hash is direct from the XML.

It also brings in and builds the Area structs using the appropriate
builder function.